laait_backend/requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.1
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.3
django-extensions==3.1.5
djangorestframework==3.13.1
duplicity==0.8.12.0
entrypoints==0.3
executing==0.8.3
fasteners==0.14.1
future==0.18.2
greenlet==0.4.15
httplib2==0.14.0
idna==2.8
ipykernel==6.9.1
ipython==8.1.1
jedi==0.18.1
jupyter-client==7.1.2
jupyter-core==4.9.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib-inline==0.1.3
monotonic==1.5
msgpack==0.6.2
nest-asyncio==1.5.4
netifaces==0.10.4
numpy==1.22.3
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.28
protobuf==3.6.1
ptpython==3.0.20
pure-eval==0.2.2
pycairo==1.16.2
pycups==1.9.73
Pygments==2.11.2
PyGObject==3.40.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pynvim==0.4.1
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.2
stack-data==0.2.0
systemd-python==234
tornado==6.1
traitlets==5.1.1
typing-extensions==4.1.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
wcwidth==0.2.5
xkit==0.0.0
